requirements.txt
torch
transformers
huggingface_hub
Pillow
numpy
streamlit